[flow] move other helpers to utils module
